+++ iOS12-HW17-Alexey-Cherebayev/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
