+++ iOS12-HW17-Alexey-Cherebayev/Resources/Assets.xcassets/Contents.json
