+++ iOS12-HW17-Alexey-Cherebayev/Resources/Info.plist
