+++ iOS12-HW17-Alexey-Cherebayev/Sources/Controllers/ViewController.swift
@@ -1,0 +1,39 @@
+//
+//  ViewController.swift
+//  iOS12-HW17-Alexey-Cherebayev
+//
+//  Created by  Alexey on 21.03.2024.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    // MARK: - UI
+    
+    // MARK: - Properties
+    
+    // MARK: - Lifecycle
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+    
+    // MARK: - Setups
+    private func setupView() {
+        
+    }
+    
+    private func setupHierarchy() {
+        
+    }
+    
+    private func setupLayout() {
+        
+    }
+
+    // MARK: - Actions
+    
+    // MARK: - Functions
+
+}
+
