--- iOS12-HW17-Alexey-Cherebayev/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  iOS12-HW17-Alexey-Cherebayev
-//
-//  Created by  Alexey on 21.03.2024.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
